Implemented priority functionality

+++ client/src/pages/Dashboard.jsx
@@ -1,6 +1,11 @@
 import React, { useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom';
 import { getAllTask, createTask, deleteTask, updateTask} from '../services/taskService';
+import { toast } from "react-toastify";
+import TaskList from '../components/TaskList';
+import TaskForm from '../components/TaskForm';
+import SearchFilter from '../components/SearchFilter';
+import DashboardStats from '../components/DashboardStats';
 
 const Dashboard = () => {
   const navigate = useNavigate()
@@ -10,10 +15,16 @@ const Dashboard = () => {
   const [formData, setFormData] = useState({
     title:"",
     description:"",
-    status:"Pending"
+    status:"Pending",
+    priority:"medium"
   })
 
   const [editTaskId, setEditTaskId] =  useState(null)
+
+  const [loading, setLoading] = useState(false)
+
+  const [searchTerm, setSearchTerm] = useState("")
+  const [statusFilter,setStatusFilter] = useState("All")
 
   const fetchTask = async () =>{
     try {
@@ -45,36 +56,46 @@ const Dashboard = () => {
   const handleSubmit = async (e) =>{
     e.preventDefault()
     try {
-      // const response = await createTask(formData)
+      setLoading(true)
       
       if(editTaskId){
         await updateTask(editTaskId, formData) 
+        toast.success("Task updated successfully!");
        }
        else{
        await createTask(formData)
+       toast.success("Task created successfully!");
        }
-
       await fetchTask();
-      // console.log(response)
-
       setFormData({
-      title:"",
-      description:"",
-      status:"Pending"
+          title:"",
+          description:"",
+          status:"Pending",
+          priority:"medium"
     })
       setEditTaskId(null)  
+      
     } catch (error) {
       console.error(error)
+      toast.error("Failed to save task!");
+    }finally{
+      setLoading(false)
     }
   }
 
   const handleDelete = async (taskId) =>{
     try {
+      const isConfirmed = window.confirm("Are you sure you want to delete this task ?")
+      if(!isConfirmed) {
+        return
+      }
       const response = await deleteTask(taskId)
+      toast.success("Task deleted successfully!");
 
-        await fetchTask();
+      await fetchTask();
     } catch (error) {
       console.error(error)
+      toast.error("Failed to delete task!");
     }
   }
 
@@ -86,81 +107,51 @@ const Dashboard = () => {
           title: task.title,
           description: task.description,
           status: task.status,
+          priority: task.priority
         });
-      // const response = await updateTask()
-
-      // await fetchTask()
     
    } catch (error) {
        console.error(error)
     }
   }
 
-  console.log(formData)
+     const filteredTasks = tasks.filter((task) => {
+      const matchesSearch = task.title
+        .toLowerCase()
+        .includes(searchTerm.toLowerCase());
+
+  const matchesStatus =
+    statusFilter === "All" || task.status === statusFilter;
+
+  return matchesSearch && matchesStatus;
+});
   return (
     <div>
 
-      {/* <button className='bg-blue-600 text-white px-4 py-2 rounded mb-4'>Add Task</button> */}
       <button onClick={handleLogout}
       className='bg-red-500 text-white px-4 py-2 rounded' >
         Logout
       </button>
 
+      <DashboardStats tasks={tasks}/>
+      <SearchFilter
+        searchTerm={searchTerm}
+        setSearchTerm={setSearchTerm}
+        statusFilter={statusFilter}
+        setStatusFilter={setStatusFilter} />
 
-      <form onSubmit={handleSubmit} className='mt-6 space-y-4'>
-        <input type="text" 
-        placeholder='Enter Task Title'
-        name='title'
-        value={formData.title} 
-        onChange={handleChange}
-        className='w-full rounded-lg border p-2' />
+        <TaskForm  handleSubmit={handleSubmit}
+          formData={formData}
+          handleChange={handleChange}
+          loading={loading}
+          editTaskId={editTaskId}/>
 
-        <textarea placeholder='Enter Task Description' 
-        className='w-full rounded-lg border p-2'
-        name='description'
-        value={formData.description}
-        onChange={handleChange}
-        rows="4"></textarea>
+        <TaskList
+          tasks={filteredTasks}
+          onEdit={handleEdit}
+          onDelete={handleDelete}/>
 
-        <select className='w-full rounded-lg border p-2'
-        name='status'
-        value={formData.status}
-        onChange={handleChange}>
-          <option value="Pending">Pending</option>
-          <option value="In Progress">In Progress</option>
-          <option value="Completed">Completed</option>
-        </select>
-
-        <button type='submit'
-        className='bg-blue-600 rounded-lg text-white px-4 py-2'> {editTaskId ? "Update Task" : "Add Task"}</button>
-      </form>
-
-      {tasks.map((task)=> {
-        return  (
-        <div key={task._id}
-        className='border rounded-lg p-4 shadow-md mt-4'>
-           <h2 className='text-xl font-bold'>{task.title}</h2>
-           <p>{task.description}</p>
-           <p>Status :{task.status}</p>
-
-           <div className="mt-3 flex gap-2">
-              <button
-                onClick={() => handleEdit(task)}
-                className="bg-blue-500 hover:bg-blue-600 text-white px-3 py-2 rounded"
-              >
-                Edit
-              </button>
-
-              <button
-                onClick={() => handleDelete(task._id)}
-                className="bg-red-500 hover:bg-red-600 text-white px-3 py-2 rounded"
-              >
-               Delete
-              </button>
-            </div>
-        </div> )
-      })}
-    </div>
+     </div>
   )
 }
